LayoutTests/editing/selection/ios/show-selection-in-transformed-container-2-expected.txt


This test verifies that after focusing a visible input field in a display: flex container that has been translated horizontally out of view, the caret is still visible. To run the test manually, tap the input field and check that the caret shows up.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL caretRect.left should be 10. Was 9.
PASS caretRect.top is 4
FAIL caretRect.width should be 2. Was 3.
PASS caretRect.height is 24
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
